requirements.txt
python-telegram-bot
psycopg2-binary
SQLAlchemy
openai>=1.0.0
fastapi
uvicorn
